training/requirements.txt
torch==2.6.0
torchvision==0.21.0
torchmetrics==1.4.0
torchsampler==0.1.2
lightning==2.5.0.post0
tensorboard==2.19.0
clearml==2.0.2
timm==1.0.19


hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

scikit-learn==1.6.1
scikit-image==0.25.2
opencv-python==4.9.0.80
numpy==1.26.4
pandas==2.2.3
openpyxl==3.1.5

rootutils==1.0.5
requests==2.32.3
tqdm==4.67.1
pillow==11.0.0

rich==13.9.4
scipy==1.12.0
kaggle==1.7.4.5
